+++ HCM_Project/src/main/java/com/hcm/grw/model/mapper/ChatDao.java
@@ -10,8 +10,8 @@ public interface ChatDao {
 	
 	public List<ChatDto> selectAllMessage(ChatDto dto);
 	
-	public int setReadMessage(String ch_target);
+	public int setReadMessage(ChatDto dto);
 	
-	public int noReadList(String ch_target);
+	public int noReadList(ChatDto dto);
 	
 }
