+++ HCM_Project/src/main/java/com/hcm/grw/model/service/ChatServiceImpl.java
@@ -30,15 +30,15 @@ public class ChatServiceImpl implements ChatService {
 	}
 	
 	@Override
-	public int setReadMessage(String ch_target) {
+	public int setReadMessage(ChatDto dto) {
 		log.info("ChatServiceImpl setReadMessage Service 읽음처리");
-		return dao.setReadMessage(ch_target);
+		return dao.setReadMessage(dto);
 	}
 	
 	@Override
-	public int noReadList(String ch_target) {
+	public int noReadList(ChatDto dto) {
 		log.info("ChatServiceImpl noReadList Service 안읽은 목록 수");
-		return dao.noReadList(ch_target);
+		return dao.noReadList(dto);
 	}
 
 }
